feat: Add `FunctionExpr` and `EmptyListExpr`
These two functions produce expressions that are easier to write than
writing out the struct manually. The `FunctionExpr` function takes an
interface because I can't/don't want to predict the function signature
of the argument.

Perhaps I will change my mind and add a function signature when the API
is stable.